Nowiki blocks work
